requirements.txt
joblib
scikit-learn
Pillow
numpy